requirements.txt
git+https://github.com/Rapptz/discord.py
git+https://github.com/Rapptz/discord-ext-menus
discord-flags==2.1.1
async-timeout==3.0.1
DateTime==4.3
aiohttp==3.6.2
cffi==1.14.3
PyNaCl==1.3.0
pycparser==2.20
yarl==1.4.2
jishaku @ git+https://github.com/Gorialis/jishaku@98a4e0f2f186d961231db20b61cee80450f69fb9
toml==0.9.0